src/requirements.txt
requests
openai
fastapi
uvicorn
python-docx
